fix: factoring out utils and tasks.py
