+++ enterprise_access/apps/api/utils.py
@@ -6,9 +6,6 @@ from uuid import UUID
 
 from rest_framework.exceptions import ParseError
 
-from enterprise_access.apps.subsidy_request.constants import SubsidyTypeChoices
-from enterprise_access.apps.subsidy_request.models import CouponCodeRequest, LicenseRequest
-
 
 def get_enterprise_uuid_from_query_params(request):
     """
@@ -39,21 +36,6 @@ def get_enterprise_uuid_from_request_data(request):
         return UUID(enterprise_customer_uuid)
     except ValueError as ex:
         raise ParseError('{} is not a valid uuid.'.format(enterprise_customer_uuid)) from ex
-
-
-def get_subsidy_model(subsidy_type):
-    """
-    Get subsidy model from subsidy_type string
-
-    Args:
-        subsidy_type (string): string name of subsidy
-    """
-    subsidy_model = None
-    if subsidy_type == SubsidyTypeChoices.COUPON:
-        subsidy_model = CouponCodeRequest
-    if subsidy_type == SubsidyTypeChoices.LICENSE:
-        subsidy_model = LicenseRequest
-    return subsidy_model
 
 
 # Can use this to replace above logic in other utils functions,
