+++ enterprise_access/utils.py
@@ -1,0 +1,23 @@
+"""
+Utils for any app in the enterprise-access project.
+"""
+
+from django.apps import apps
+from enterprise_access.apps.subsidy_request.constants import SubsidyTypeChoices
+
+
+def get_subsidy_model(subsidy_type):
+    """
+    Get subsidy model from subsidy_type string
+
+    Args:
+        subsidy_type (string): string name of subsidy
+    Returns:
+        Class of a model object
+    """
+    subsidy_model = None
+    if subsidy_type == SubsidyTypeChoices.COUPON:
+        subsidy_model = apps.get_model('subsidy_request.CouponCodeRequest')
+    if subsidy_type == SubsidyTypeChoices.LICENSE:
+        subsidy_model = apps.get_model('subsidy_request.LicenseRequest')
+    return subsidy_model
